Time: 0 ms (100.00%) | Memory: 9.5 MB (98.46%) - LeetSync

+++ 2812-find-the-maximum-achievable-number/find-the-maximum-achievable-number.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int theMaximumAchievableX(int num, int t) {
+    short theMaximumAchievableX(short num, short t) {
         return num + t + t;
     }
 };
